Corretti bug, aggiunta la possibilità di vedere i vecchi viaggi dal rider. Quando si inserisce l'indirizzo di consegna automaticamente, l'app in automatico restituisce i valori esatti.

+++ app/src/main/java/it/benche/upofood/RequestsActivity.kt
@@ -79,7 +79,7 @@ class RequestsActivity: AppCompatActivity() {
                                         .get()
                                         .addOnSuccessListener { document ->
                                             val user = document.getString("client").toString()
-                                            val price = document.getLong("totalPrice")!!.toString()
+                                            val price = document.getDouble("totalPrice")!!.toString()
 
                                             db.collection("users")
                                                 .document(user)
